requirements.txt
torch
torchvision
torchaudio
streamlit
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
matplotlib
seaborn
numpy
faiss-cpu
joblib
sentence-transformers